Add ability to request reviewers when creating a PR (#40)
This adds a new section to the PR creation modal that allows you to
select reviewers. It'll try to pull the list of collaborators on your
current repo to auto-populate suggestions, but there is also a new
configuration option that will let you set "additional reviewers" that
will always be suggested. These are useful if you have a common set of
folks that you work with and want to use reference them as reviewers :)

---------

Co-authored-by: gizmo385 <gizmo385@users.noreply.github.com>